+++ website/site/data/spec.js
@@ -1,5 +1,5 @@
 window.__FLIP_SPEC__ = {
-  "generated": "2026-08-18T02:57:38+00:00",
+  "generated": "2026-08-18T03:06:53+00:00",
   "sections": [
     {
       "number": "1",